+++ jcl/jpp_configuration.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE configurationreg SYSTEM "http://home.ottawa.ibm.com/teams/bluebird/web/eclipse_site/jpp.dtd">
 <!--
 /*******************************************************************************
- * Copyright (c) 2002, 2019 IBM Corp. and others
+ * Copyright (c) 2002, 2020 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -82,6 +82,7 @@
 		  dependencies="SIDECAR18-SE"
 		  jdkcompliance="1.8">
 		<classpathentry kind="lib" path="/binaries/vm/third/rt-compressed.sun180.jar"/>
+		<source path="src/com.ibm.jzos/share/classes"/>
 		<source path="src/java.base/share/classes"/>
 		<source path="src/java.desktop/share/classes"/>
 		<source path="src/java.logging/share/classes"/>
@@ -106,6 +107,7 @@
 		  flags="Sidecar18-SE-OpenJ9, DAA, Open-Module-Support,Sidecar19-SE,Sidecar19-SE-OpenJ9,Java10,Java11"
 		  dependencies="SIDECAR18-SE"
 		  jdkcompliance="1.8">
+		<classpathentry kind="src" path="src/com.ibm.jzos/share/classes"/>
 		<classpathentry kind="src" path="src/java.base/share/classes"/>
 		<classpathentry kind="src" path="src/java.management/share/classes"/>
 		<classpathentry kind="src" path="src/jdk.attach/share/classes"/>
@@ -134,6 +136,7 @@
 		  flags="Java12,Java13"
 		  dependencies="JAVA11"
 		  jdkcompliance="1.8">
+		<classpathentry kind="src" path="src/com.ibm.jzos/share/classes"/>
 		<classpathentry kind="src" path="src/java.base/share/classes"/>
 		<classpathentry kind="src" path="src/java.management/share/classes"/>
 		<classpathentry kind="src" path="src/jdk.attach/share/classes"/>
@@ -162,6 +165,7 @@
 		  flags="Java14"
 		  dependencies="JAVA13"
 		  jdkcompliance="1.8">
+		<classpathentry kind="src" path="src/com.ibm.jzos/share/classes"/>
 		<classpathentry kind="src" path="src/java.base/share/classes"/>
 		<classpathentry kind="src" path="src/java.management/share/classes"/>
 		<classpathentry kind="src" path="src/jdk.attach/share/classes"/>
@@ -190,6 +194,7 @@
 		  flags="Java15"
 		  dependencies="JAVA14"
 		  jdkcompliance="1.8">
+		<classpathentry kind="src" path="src/com.ibm.jzos/share/classes"/>
 		<classpathentry kind="src" path="src/java.base/share/classes"/>
 		<classpathentry kind="src" path="src/java.management/share/classes"/>
 		<classpathentry kind="src" path="src/jdk.attach/share/classes"/>
@@ -218,6 +223,7 @@
 		  flags="OpenJ9-RawBuild"
 		  dependencies="JAVA11"
 		  jdkcompliance="1.8">
+		<classpathentry kind="src" path="src/com.ibm.jzos/share/classes"/>
 		<classpathentry kind="src" path="src/java.base/share/classes"/>
 		<classpathentry kind="src" path="src/java.management/share/classes"/>
 		<classpathentry kind="src" path="src/jdk.attach/share/classes"/>
@@ -246,6 +252,7 @@
 		  flags="Panama"
 		  dependencies="JAVA13"
 		  jdkcompliance="1.8">
+		<classpathentry kind="src" path="src/com.ibm.jzos/share/classes"/>
 		<classpathentry kind="src" path="src/java.base/share/classes"/>
 		<classpathentry kind="src" path="src/java.management/share/classes"/>
 		<classpathentry kind="src" path="src/jdk.attach/share/classes"/>
